move around getActivityList and update comments.
