+++ frontend/src/components/ViewActivities/activityfns.js
@@ -4,37 +4,6 @@ import app from '../Firebase';
 const db = app.firestore();
 
 /**
- * Gets the list of activities from the server. 
- * @param {string} tripId The trip ID.
- */
-export async function getActivityList(tripId) {
-  return new Promise(function(resolve, reject) {
-    let tripActivities = [];
-    
-    db.collection(DBFIELDS.COLLECTION_TRIPS).doc(tripId)
-    .collection(DBFIELDS.COLLECTION_ACTIVITIES).get()
-    .then(querySnapshot => {
-      querySnapshot.forEach(doc => {
-        let data = doc.data();
-        data['id'] = doc.id;
-        
-        // TODO: if start date != end date, split into 2 days. (#37)
-
-        // Eliminate nanoseconds, convert to milliseconds.
-        data[DBFIELDS.ACTIVITIES_START_TIME] =
-          data[DBFIELDS.ACTIVITIES_START_TIME]['seconds'] * 1000;         
-        data[DBFIELDS.ACTIVITIES_END_TIME] = 
-          data[DBFIELDS.ACTIVITIES_END_TIME]['seconds'] * 1000;
-
-        tripActivities.push(data);
-      })
-    }).catch(error => {
-      tripActivities = null;
-    }).then( () => resolve(tripActivities) );
-  })
-}
-
-/**
  * Sort a list of trip activities by date. 
  * @param {Array} tripActivities Array of activities.
  * @returns List of trip activities in the form
