+++ frontend/src/components/ViewActivities/activityfns.test.js
@@ -1,4 +1,5 @@
 import * as activityFns from './activityfns.js';
+import { getActivityList } from './activitylist.js';
 
 const ten = new Date(Date.UTC(2020, 4, 2, 10,  0));          // May 2, 2020 10:00
 const eleven = new Date(Date.UTC(2020, 4, 2, 11, 0));        // May 2, 2020 11:00
